refactor(time) : 가독성을 위해 time에 대한 영역으로 판단할 수 있도록 time 디렉토리로 네이밍 변경
