+++ src/test/java/com/aliens/backend/matching/time/MockTime.java
@@ -1,4 +1,4 @@
-package com.aliens.backend.matching.util;
+package com.aliens.backend.matching.time;
 
 import java.time.LocalDateTime;
